+++ src/backend/TrafficCourts/Workflow.Service/Configuration/OracleDataApiConfiguration.cs
@@ -1,0 +1,8 @@
+﻿namespace TrafficCourts.Workflow.Service.Configuration
+{
+    public class OracleDataApiConfiguration
+    {
+        public string? Host { get; set; }
+        public int Port { get; set; }
+    }
+}
