--- src/backend/TrafficCourts/Workflow.Service/Configuration/OracleInterfaceApiConfiguration.cs
@@ -1,8 +1,0 @@
-﻿namespace TrafficCourts.Workflow.Service.Configuration
-{
-    public class OracleInterfaceApiConfiguration
-    {
-        public string? Host { get; set; }
-        public int Port { get; set; }
-    }
-}
