+++ src/backend/TrafficCourts/Workflow.Service/Services/IOracleDataApiService.cs
@@ -1,0 +1,9 @@
+﻿using TrafficCourts.Workflow.Service.Models;
+
+namespace TrafficCourts.Workflow.Service.Services
+{
+    public interface IOracleDataApiService
+    {
+        Task<int> CreateDisputeAsync(Dispute disputeToSubmit);
+    }
+}
