--- src/backend/TrafficCourts/Workflow.Service/Services/IOracleInterfaceService.cs
@@ -1,9 +1,0 @@
-﻿using TrafficCourts.Workflow.Service.Models;
-
-namespace TrafficCourts.Workflow.Service.Services
-{
-    public interface IOracleInterfaceService
-    {
-        Task<int> CreateDisputeAsync(Dispute disputeToSubmit);
-    }
-}
