+++ src/backend/TrafficCourts/Workflow.Service/appsettings.json
@@ -13,7 +13,7 @@
     "Host": "localhost",
     "Port": "5300"
   },
-  "OracleInterfaceApiConfiguration": {
+  "OracleDataApiConfiguration": {
     "Host": "localhost",
     "Port": "5010"
   },
