+++ src/backend/oracle-data-api/.env.template
