+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/repository/DisputeRepository.java
@@ -1,0 +1,9 @@
+package ca.bc.gov.open.jag.tco.oracledataapi.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import ca.bc.gov.open.jag.tco.oracledataapi.model.Dispute;
+
+public interface DisputeRepository extends CrudRepository<Dispute, Integer>{
+
+}
