+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/repository/StatuteRepository.java
@@ -1,0 +1,9 @@
+package ca.bc.gov.open.jag.tco.oracledataapi.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import ca.bc.gov.open.jag.tco.oracledataapi.model.Statute;
+
+public interface StatuteRepository extends CrudRepository<Statute, Integer>{
+
+}
