+++ src/backend/oracle-data-api/src/test/java/ca/bc/gov/open/jag/tco/oracledataapi/OracleDataApiApplicationTests.java
@@ -1,0 +1,13 @@
+package ca.bc.gov.open.jag.tco.oracledataapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OracleDataApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
