--- src/backend/oracle-data-interface/src/main/java/ca/bc/gov/open/jag/tco/oracledatainterface/repository/StatuteRepository.java
@@ -1,9 +1,0 @@
-package ca.bc.gov.open.jag.tco.oracledatainterface.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import ca.bc.gov.open.jag.tco.oracledatainterface.model.Statute;
-
-public interface StatuteRepository extends CrudRepository<Statute, Integer>{
-
-}
