--- src/backend/oracle-data-interface/src/test/java/ca/bc/gov/open/jag/tco/oracledatainterface/OracleDataInterfaceApplicationTests.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.jag.tco.oracledatainterface;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OracleDataInterfaceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
